+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to
 - 🚀(front) fix frontend build failure
 - 🐛(makefile) fix args in make test
 - 🩹(backend) fix case-insensitive email deduplication in merge command
+- 🐛(summary) support media files with bad streams #1478
 
 ## [1.22.0] - 2026-07-03